Added USB install menu
Added install menu that allows users to select what to install on the USB instead of just the full MediCat package.

This should be simple to implement and test.. but also you can create your own combo of items to install on the USB that might be better than this combo.